PHPUnit tests for factory
