+++ psalm.xml
@@ -12,7 +12,7 @@
             <directory name="src/Bundle/spec" />
             <directory name="src/Component/spec" />
             <directory name="src/Component/Tests" />
-            <directory name="src/Component/tests/spec" />
+            <directory name="src/Component/tests" />
             <directory name="src/Component/vendor" />
             <file name="src/Bundle/Controller/ControllerTrait.php" />
             <file name="src/Bundle/DependencyInjection/Driver/Doctrine/DoctrineODMDriver.php" />
@@ -165,9 +165,6 @@
 
         <PropertyNotSetInConstructor>
             <errorLevel type="suppress">
-                <file name="src/Component/Tests/Reflection/ClassInfoTraitTest.php" />
-                <file name="src/Component/Tests/Reflection/ClassReflectionTest.php" />
-                <file name="src/Component/Tests/Symfony/ExpressionLanguage/ArgumentParserTest.php" />
                 <file name="src/Bundle/Validator/Constraints/UniqueWithinCollectionConstraint.php" />
                 <file name="src/Bundle/Validator/Constraints/Enabled.php" />
                 <file name="src/Bundle/Validator/Constraints/Disabled.php" />
