+++ src/Component/tests/Symfony/EventListener/DeserializeListenerTest.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Resource\Tests\Symfony\EventListener;
+namespace Sylius\Resource\Tests\Symfony\EventListener;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
