+++ src/Component/tests/Symfony/EventListener/FactoryListenerTest.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Resource\Tests\Symfony\EventListener;
+namespace Sylius\Resource\Tests\Symfony\EventListener;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
