+++ src/Component/tests/Symfony/EventListener/FlashListenerTest.php
@@ -11,38 +11,45 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Resource\Tests\State\Processor;
+namespace Sylius\Resource\Tests\Symfony\EventListener;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
 use Sylius\Resource\Context\Context;
-use Sylius\Resource\Context\Option\RequestOption;
+use Sylius\Resource\Context\Initiator\RequestContextInitiatorInterface;
 use Sylius\Resource\Metadata\HttpOperation;
-use Sylius\Resource\State\Processor\FlashProcessor;
-use Sylius\Resource\State\ProcessorInterface;
+use Sylius\Resource\Metadata\Operation\HttpOperationInitiatorInterface;
+use Sylius\Resource\Symfony\EventListener\FlashListener;
 use Sylius\Resource\Symfony\Session\Flash\FlashHelperInterface;
 use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Event\ViewEvent;
+use Symfony\Component\HttpKernel\HttpKernelInterface;
+use Symfony\Component\HttpKernel\KernelInterface;
 
-final class FlashProcessorTest extends TestCase
+final class FlashListenerTest extends TestCase
 {
     use ProphecyTrait;
 
-    private ProcessorInterface|ObjectProphecy $decorated;
+    private HttpOperationInitiatorInterface|ObjectProphecy $operationInitiator;
+
+    private RequestContextInitiatorInterface|ObjectProphecy $requestContextInitiator;
 
     private FlashHelperInterface|ObjectProphecy $flashHelper;
 
-    private FlashProcessor $flashProcessor;
+    private FlashListener $flashListener;
 
     protected function setUp(): void
     {
-        $this->decorated = $this->prophesize(ProcessorInterface::class);
+        $this->operationInitiator = $this->prophesize(HttpOperationInitiatorInterface::class);
+        $this->requestContextInitiator = $this->prophesize(RequestContextInitiatorInterface::class);
         $this->flashHelper = $this->prophesize(FlashHelperInterface::class);
 
-        $this->flashProcessor = new FlashProcessor(
-            $this->decorated->reveal(),
+        $this->flashListener = new FlashListener(
+            $this->operationInitiator->reveal(),
+            $this->requestContextInitiator->reveal(),
             $this->flashHelper->reveal(),
         );
     }
@@ -50,9 +57,17 @@ final class FlashProcessorTest extends TestCase
     /** @test */
     public function it_adds_flash(): void
     {
+        $kernel = $this->prophesize(KernelInterface::class);
         $request = $this->prophesize(Request::class);
         $attributes = $this->prophesize(ParameterBag::class);
         $operation = $this->prophesize(HttpOperation::class);
+
+        $event = new ViewEvent(
+            $kernel->reveal(),
+            $request->reveal(),
+            HttpKernelInterface::MAIN_REQUEST,
+            ['foo' => 'fighters'],
+        );
 
         $request->isMethodSafe()->willReturn(false);
 
@@ -60,22 +75,32 @@ final class FlashProcessorTest extends TestCase
 
         $attributes->getBoolean('is_valid', true)->willReturn(true);
 
-        $context = new Context(new RequestOption($request->reveal()));
+        $this->operationInitiator->initializeOperation($request)->willReturn($operation);
 
-        $this->decorated->process(['foo' => 'fighters'], $operation, $context)->willReturn(['foo' => 'fighters'])->shouldBeCalled();
+        $context = new Context();
+
+        $this->requestContextInitiator->initializeContext($request)->willReturn($context);
 
         $this->flashHelper->addSuccessFlash($operation, $context)->shouldBeCalled();
 
-        $this->flashProcessor->process(['foo' => 'fighters'], $operation->reveal(), $context);
+        $this->flashListener->onKernelView($event);
     }
 
     /** @test */
     public function it_does_nothing_when_controller_result_is_a_response(): void
     {
+        $kernel = $this->prophesize(KernelInterface::class);
         $request = $this->prophesize(Request::class);
         $attributes = $this->prophesize(ParameterBag::class);
         $operation = $this->prophesize(HttpOperation::class);
         $response = $this->prophesize(Response::class);
+
+        $event = new ViewEvent(
+            $kernel->reveal(),
+            $request->reveal(),
+            HttpKernelInterface::MAIN_REQUEST,
+            $response->reveal(),
+        );
 
         $request->isMethodSafe()->willReturn(false)->shouldNotBeCalled();
 
@@ -83,21 +108,31 @@ final class FlashProcessorTest extends TestCase
 
         $attributes->getBoolean('is_valid', true)->willReturn(true)->shouldNotBeCalled();
 
-        $context = new Context(new RequestOption($request->reveal()));
+        $this->operationInitiator->initializeOperation($request)->willReturn($operation);
 
-        $this->decorated->process($response, $operation, $context)->willReturn($response)->shouldBeCalled();
+        $context = new Context();
+
+        $this->requestContextInitiator->initializeContext($request)->willReturn($context);
 
         $this->flashHelper->addSuccessFlash($operation, $context)->shouldNotBeCalled();
 
-        $this->flashProcessor->process($response->reveal(), $operation->reveal(), $context);
+        $this->flashListener->onKernelView($event);
     }
 
     /** @test */
     public function it_does_nothing_when_method_is_safe(): void
     {
+        $kernel = $this->prophesize(KernelInterface::class);
         $request = $this->prophesize(Request::class);
         $attributes = $this->prophesize(ParameterBag::class);
         $operation = $this->prophesize(HttpOperation::class);
+
+        $event = new ViewEvent(
+            $kernel->reveal(),
+            $request->reveal(),
+            HttpKernelInterface::MAIN_REQUEST,
+            ['foo' => 'fighters'],
+        );
 
         $request->isMethodSafe()->willReturn(true)->shouldBeCalled();
 
@@ -105,21 +140,31 @@ final class FlashProcessorTest extends TestCase
 
         $attributes->getBoolean('is_valid', true)->willReturn(true);
 
-        $context = new Context(new RequestOption($request->reveal()));
+        $this->operationInitiator->initializeOperation($request)->willReturn($operation);
 
-        $this->decorated->process(['foo' => 'fighters'], $operation, $context)->willReturn(['foo' => 'fighters'])->shouldBeCalled();
+        $context = new Context();
+
+        $this->requestContextInitiator->initializeContext($request)->willReturn($context);
 
         $this->flashHelper->addSuccessFlash($operation, $context)->shouldNotBeCalled();
 
-        $this->flashProcessor->process(['foo' => 'fighters'], $operation->reveal(), $context);
+        $this->flashListener->onKernelView($event);
     }
 
     /** @test */
     public function it_does_nothing_when_validation_has_failed(): void
     {
+        $kernel = $this->prophesize(KernelInterface::class);
         $request = $this->prophesize(Request::class);
         $attributes = $this->prophesize(ParameterBag::class);
         $operation = $this->prophesize(HttpOperation::class);
+
+        $event = new ViewEvent(
+            $kernel->reveal(),
+            $request->reveal(),
+            HttpKernelInterface::MAIN_REQUEST,
+            ['foo' => 'fighters'],
+        );
 
         $request->isMethodSafe()->willReturn(false);
 
@@ -127,12 +172,14 @@ final class FlashProcessorTest extends TestCase
 
         $attributes->getBoolean('is_valid', true)->willReturn(false)->shouldBeCalled();
 
-        $context = new Context(new RequestOption($request->reveal()));
+        $this->operationInitiator->initializeOperation($request)->willReturn($operation);
 
-        $this->decorated->process(['foo' => 'fighters'], $operation, $context)->willReturn(['foo' => 'fighters'])->shouldBeCalled();
+        $context = new Context();
+
+        $this->requestContextInitiator->initializeContext($request)->willReturn($context);
 
         $this->flashHelper->addSuccessFlash($operation, $context)->shouldNotBeCalled();
 
-        $this->flashProcessor->process(['foo' => 'fighters'], $operation->reveal(), $context);
+        $this->flashListener->onKernelView($event);
     }
 }
