+++ src/Component/tests/Symfony/EventListener/ReadListenerTest.php
@@ -11,65 +11,64 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Resource\Tests\State\Provider;
+namespace Sylius\Resource\Tests\Symfony\EventListener;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
 use Sylius\Resource\Context\Context;
-use Sylius\Resource\Context\Option\RequestOption;
+use Sylius\Resource\Context\Initiator\RequestContextInitiatorInterface;
 use Sylius\Resource\Metadata\Create;
 use Sylius\Resource\Metadata\HttpOperation;
-use Sylius\Resource\State\Provider\ReadProvider;
+use Sylius\Resource\Metadata\Operation\HttpOperationInitiatorInterface;
 use Sylius\Resource\State\ProviderInterface;
+use Sylius\Resource\Symfony\EventListener\ReadListener;
 use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
-final class ReadProviderTest extends TestCase
+final class ReadListenerTest extends TestCase
 {
     use ProphecyTrait;
 
+    private HttpOperationInitiatorInterface|ObjectProphecy $operationInitiator;
+
+    private RequestContextInitiatorInterface|ObjectProphecy $contextInitiator;
+
     private ProviderInterface|ObjectProphecy $provider;
 
-    private ReadProvider $readProvider;
+    private ReadListener $readListener;
 
     protected function setUp(): void
     {
+        $this->operationInitiator = $this->prophesize(HttpOperationInitiatorInterface::class);
+        $this->contextInitiator = $this->prophesize(RequestContextInitiatorInterface::class);
         $this->provider = $this->prophesize(ProviderInterface::class);
 
-        $this->readProvider = new ReadProvider(
+        $this->readListener = new ReadListener(
+            $this->operationInitiator->reveal(),
+            $this->contextInitiator->reveal(),
             $this->provider->reveal(),
         );
     }
 
     /** @test */
-    public function it_retrieves_data(): void
+    public function it_retrieves_data_and_store_them_to_request(): void
     {
+        $event = $this->prophesize(RequestEvent::class);
         $request = $this->prophesize(Request::class);
         $attributes = $this->prophesize(ParameterBag::class);
         $operation = $this->prophesize(HttpOperation::class);
+
+        $event->getRequest()->willReturn($request);
 
         $context = new Context();
 
-        $request->attributes = $attributes;
+        $this->contextInitiator->initializeContext($request)->willReturn($context);
 
-        $this->provider->provide($operation, Argument::type(Context::class))->willReturn(['foo' => 'fighters'])->shouldBeCalled();
-
-        $attributes->set('data', Argument::any())->shouldNotBeCalled();
-
-        $this->readProvider->provide($operation->reveal(), $context);
-    }
-
-    /** @test */
-    public function it_retrieves_data_and_store_them_to_request(): void
-    {
-        $request = $this->prophesize(Request::class);
-        $attributes = $this->prophesize(ParameterBag::class);
-        $operation = $this->prophesize(HttpOperation::class);
-
-        $context = new Context(new RequestOption($request->reveal()));
+        $this->operationInitiator->initializeOperation($request)->willReturn($operation);
 
         $request->attributes = $attributes;
 
@@ -77,17 +76,25 @@ final class ReadProviderTest extends TestCase
 
         $attributes->set('data', ['foo' => 'fighters'])->shouldBeCalled();
 
-        $this->readProvider->provide($operation->reveal(), $context);
+        $this->readListener->onKernelRequest($event->reveal());
     }
 
     /** @test */
     public function it_does_nothing_when_operation_is_a_create_operation(): void
     {
+        $event = $this->prophesize(RequestEvent::class);
         $request = $this->prophesize(Request::class);
         $attributes = $this->prophesize(ParameterBag::class);
+
+        $event->getRequest()->willReturn($request);
+
         $operation = new Create();
 
-        $context = new Context(new RequestOption($request->reveal()));
+        $context = new Context();
+
+        $this->contextInitiator->initializeContext($request)->willReturn($context);
+
+        $this->operationInitiator->initializeOperation($request)->willReturn($operation);
 
         $request->attributes = $attributes;
 
@@ -95,17 +102,24 @@ final class ReadProviderTest extends TestCase
 
         $attributes->set('data', Argument::any())->shouldNotBeCalled();
 
-        $this->readProvider->provide($operation, $context);
+        $this->readListener->onKernelRequest($event->reveal());
     }
 
     /** @test */
     public function it_does_nothing_when_operation_cannot_be_read(): void
     {
+        $event = $this->prophesize(RequestEvent::class);
         $request = $this->prophesize(Request::class);
         $attributes = $this->prophesize(ParameterBag::class);
         $operation = $this->prophesize(HttpOperation::class);
 
-        $context = new Context(new RequestOption($request->reveal()));
+        $event->getRequest()->willReturn($request);
+
+        $context = new Context();
+
+        $this->contextInitiator->initializeContext($request)->willReturn($context);
+
+        $this->operationInitiator->initializeOperation($request)->willReturn($operation);
 
         $request->attributes = $attributes;
 
@@ -118,17 +132,24 @@ final class ReadProviderTest extends TestCase
 
         $attributes->set('data', Argument::any())->shouldNotBeCalled();
 
-        $this->readProvider->provide($operation->reveal(), $context);
+        $this->readListener->onKernelRequest($event->reveal());
     }
 
     /** @test */
     public function it_throws_an_exception_when_no_data_was_found(): void
     {
+        $event = $this->prophesize(RequestEvent::class);
         $request = $this->prophesize(Request::class);
         $attributes = $this->prophesize(ParameterBag::class);
         $operation = $this->prophesize(HttpOperation::class);
 
-        $context = new Context(new RequestOption($request->reveal()));
+        $event->getRequest()->willReturn($request);
+
+        $context = new Context();
+
+        $this->contextInitiator->initializeContext($request)->willReturn($context);
+
+        $this->operationInitiator->initializeOperation($request)->willReturn($operation);
 
         $request->attributes = $attributes;
 
@@ -141,6 +162,6 @@ final class ReadProviderTest extends TestCase
         $this->expectException(NotFoundHttpException::class);
         $this->expectExceptionMessage('Resource has not been found.');
 
-        $this->readProvider->provide($operation->reveal(), $context);
+        $this->readListener->onKernelRequest($event->reveal());
     }
 }
