--- src/Component/tests/spec/Factory/FactorySpec.php
@@ -1,35 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Sylius Sp. z o.o.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace spec\Sylius\Resource\Factory;
-
-use PhpSpec\ObjectBehavior;
-use Sylius\Resource\Factory\FactoryInterface;
-
-final class FactorySpec extends ObjectBehavior
-{
-    function let(): void
-    {
-        $this->beConstructedWith(\stdClass::class);
-    }
-
-    function it_implements_factory_interface(): void
-    {
-        $this->shouldHaveType(FactoryInterface::class);
-    }
-
-    function it_creates_a_new_instance_of_a_resource(): void
-    {
-        $this->createNew()->shouldHaveType(\stdClass::class);
-    }
-}
